fix dữ liệu đồng bộ theo dulieunguoichoi.java thay vì nửa nhanvat.java & dulieunguoichoi.java + fix nhabroly.java cho hành tinh xayda

+++ core/src/main/java/com/dang/dragonboy/du_lieu/DuLieuNguoiChoi.java
@@ -194,7 +194,7 @@ public class DuLieuNguoiChoi {
         return null;
     }
 
-    public void tangSucManh(int SucManhCongThem){
+    public void tangSucManh(long SucManhCongThem){
         this.sucManh += SucManhCongThem;
     }
